requirements.txt
fastapi
anthropic
pydantic
python-dotenv
